Moving part of logic from TripController to PointController
